HTTP Status included and moved to Public API
