+++ platform/dolphin-platform-client/src/main/java/com/canoo/platform/client/HttpStatus.java
@@ -13,13 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dp.impl.platform.client.http;
+package com.canoo.platform.client;
 
 public interface HttpStatus {
 
-    int HTTP_OK = 200;
+	int HTTP_OK = 200;
 
-    int SC_REQUEST_TIMEOUT = 408;
+	int SC_REQUEST_TIMEOUT = 408;
 
-    int SC_MULTIPLE_CHOICES = 300;
+	int SC_MULTIPLE_CHOICES = 300;
+
+	int SC_MOVED_PERMANENTLY = 301;
+
+	int SC_FOUND = 302;
+
+	int SC_NOT_MODIFIED = 304;
+
+	int SC_TEMPORARY_REDIRECT = 307;
+
+	int SC_BAD_REQUEST = 400;
+
+	int SC_HTTP_UNAUTHORIZED = 401;
+
+	int SC_HTTP_FORBIDDEN = 403;
+
+	int SC_HTTP_RESOURCE_NOTFOUND = 404;
+
+	int SC_GONE = 410;
+
+	int SC_INTERNAL_SERVER_ERROR = 500;
+
+	int SC_NOT_IMPLMENTED = 501;
+
+	int SC_SERVICE_UNAVAILABLE = 503;
+
+	int SC_PERMISSION_DENIED = 550;
 }
